[FIX] base: correctly detect when admin logs in
Since ef79fc854f15759a2c36a1a58aa86a4212d344bc it's not possible to log in as user 1.

However, we reset the base url when the admin logs in, which is now user 2.

This fixes wkhtml2pdf was not able to load css file, and probably some other side effects.

PR: none
Task: 1879620