[IMP] mrp: BoM and product as readonly when a MO is set on UB.
Since commit ba41c80328f5072f83b65f6f4f53e1a3f11d1374 when the user specify a MO the
unbuild use the existing moves raw and moves finished in order to create the unbuild
moves. In this case the bom_id and the product_id is not needed since we don't generate
unbuild move from the BoM

This commit is related to task ID 39843.