[IMP] web: Close window when printing a report
Up until now, the execution of a report action could not be followed by
any action which prevented, for example, to automaticcaly close a modal
after printing a report from it. This has become an issue when working
on customer's invoices using the new modal introduced in fa8f20482e38fa00cbd34ac9a8426fc8f499e03a

This commit allows to close a modal after the download of a report.